+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add `EveOnline::ESI::Models::CharacterOrder`
 * Add `EveOnline::ESI::CharacterOrders`
 * Remove `EveOnline::XML::CharacterMarketOrders`. Use `EveOnline::ESI::CharacterOrders` instead.
+* Add `EveOnline::ESI::Models::CorporationOrder`
 
 **v0.14.0**